requirements.txt
appdirs==1.4.4
pygame==2.6.1
PyYAML==6.0.3
